Fix PostCSS build error if output file does not already exist
The Tailwind plugin's internals assumes that the file specified in the
`from` option in the PostCSS config exists and tries to `stat` it. This
failed on first build because we were passing the output CSS path for
this option.

We don't have an input CSS file, because the CSS is generated from SASS
inputs and only exists in memory, but we do have an input .scss file
path, so use that instead.